Add CI for image and test
This adds an image generation CI and mvn build.
The image is only built if tests are passed and on main branch.
It is based on the fairly new Chainguard distroless project,
which promises small, secure base images with SBOMS available
for license compliance.

To cleanly get there:
dependencies were bumped
maven wrapper was added
dev instructions moved to sep standard file

Signed-off-by: Josef Andersson <josef.andersson@digg.se>